[Refactor] Valores id alterados para Integer

+++ backend/src/main/java/com/example/tracker/repository/HabilidadeRepository.java
@@ -6,7 +6,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import com.example.tracker.entity.Habilidade;
 
-public interface HabilidadeRepository extends JpaRepository<Habilidade, Long> {
+public interface HabilidadeRepository extends JpaRepository<Habilidade, Integer> {
 
     Optional<Habilidade> findByDescricao(String descricao);
